+++ poem-openapi/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `content_type` attribute to the `ApiRequest` and `ApiResponse` macros to specify the content type of the request or response.
 - Panic occurs when a duplicate operation id is detected.
 - Add `OpenApiService::external_document` method to referencing an external resource for extended documentation.
+- Add `Webhook` macro to define webhooks.
+- Implement `OpenApi` for `()` to define an empty APIs.
 
 # [1.2.7] 2021-12-19